+++ AtlasLootClassic/Documentation/Release_Notes.md
@@ -1,3 +1,26 @@
+# 5.0.2 (July 24, 2025)
+- Added Dungeons and Raids + Crafting item sources
+- Added MoP Tier token data
+- Added Siege of Orgrimmar loot
+- Added Phase tags for items released in later phases
+- Fix issue where AtlasLoot didn't autoselect dungeon player is in
+- MoP factions data organized
+- Cleanup Gathering profs
+- Small fixes as usual
+
+# 5.0.1 (July 21, 2025)
+- Celestial loot organized by raid, since it's being gated
+- Challenge mode armor sets added
+- Darkmoon decks/cards and corresponding trinkets organized
+- More small fixes + locale fixes
+
+# v5.0.0 (July 19, 2025)
+- Initial release of AtlasLootClassic MoP! Lots of things are still underworks.
+- Factions, Collections, Crafting modules are mostly complete
+- Dungeons and Raids module missing SoO and Celestial dungeons
+- PvP module only displays sets right now
+- New update screen
+
 # v5.0.0-beta1 (July 1, 2025)
 - Initial MoP Pre-patch Support
 - Lots of things still broken or not populated